+++ eng/common/scripts/Verify-Links.ps1
@@ -120,6 +120,9 @@ function ProcessLink([System.Uri]$linkUri) {
     # See comment in function below for details.
     return ProcessCratesIoLink $linkUri $matches['path']
   }
+  elseif ($linkUri -match '^https?://(www\.)?npmjs\.com/package/.+') {
+    return ProcessNpmLink $linkUri
+  }
   else {
     return ProcessStandardLink $linkUri
   }
@@ -155,6 +158,14 @@ function ProcessCratesIoLink([System.Uri]$linkUri, $path) {
   Invoke-WebRequest -Uri $apiUri -Method GET -UserAgent $userAgent -TimeoutSec $requestTimeoutSec | Out-Null
   
   return $true
+}
+
+function ProcessNpmLink([System.Uri]$linkUri) {
+  # npmjs.com started using Cloudflare which returns 403 and we need to instead check the registry api for existence checks
+  # https://github.com/orgs/community/discussions/174098#discussioncomment-14461226
+  $apiUrl = $linkUri.ToString() -replace '^https?://(?:www\.)?npmjs\.com/package/(.*)/v', 'https://registry.npmjs.org/$1'
+
+  return ProcessStandardLink ([System.Uri]$apiUrl)
 }
 
 function ProcessStandardLink([System.Uri]$linkUri) {
@@ -530,6 +541,8 @@ foreach ($url in $urls) {
 
 LogGroupStart "Link checking details"
 
+$originalcheckLinkGuidance = $checkLinkGuidance
+
 while ($pageUrisToCheck.Count -ne 0)
 {
   $pageUri = $pageUrisToCheck.Dequeue();
@@ -537,6 +550,11 @@ while ($pageUrisToCheck.Count -ne 0)
   try {
     if ($checkedPages.ContainsKey($pageUri)) { continue }
     $checkedPages[$pageUri] = $true;
+
+    # copilot instructions require the use of relative links which is against our general guidance
+    # but we mainly care about those guidelines for docs publishing and not copilot instructions
+    # so we can disable the guidelines while validating copilot instruction files.
+    if ($pageUri -match "instructions.md$") { $checkLinkGuidance = $false }
 
     [string[]] $linkUris = GetLinks $pageUri
     Write-Host "Checking $($linkUris.Count) links found on page $pageUri";
@@ -561,6 +579,8 @@ while ($pageUrisToCheck.Count -ne 0)
   } catch {
     Write-Host "Exception encountered while processing pageUri $pageUri : $($_.Exception)"
     throw
+  } finally {
+    $checkLinkGuidance = $originalcheckLinkGuidance
   }
 }