+++ eng/packages/http-client-csharp-mgmt/emitter/src/sdk-context-options.ts
@@ -27,6 +27,8 @@ export const armResourceRead = "Azure.ResourceManager.@armResourceRead";
 export const armResourceReadName = "@armResourceRead";
 const armResourceReadRegex = "Azure\\.ResourceManager\\.@armResourceRead";
 
+export const readsResourceName = "@readsResource";
+
 // https://github.com/Azure/typespec-azure/blob/main/packages/typespec-azure-resource-manager/README.md#armresourcecreateorupdate
 export const armResourceCreateOrUpdate =
   "Azure.ResourceManager.@armResourceCreateOrUpdate";
@@ -53,6 +55,16 @@ const armResourceDeleteRegex = "Azure\\.ResourceManager\\.@armResourceDelete";
 export const armResourceUpdate = "Azure.ResourceManager.@armResourceUpdate";
 export const armResourceUpdateName = "@armResourceUpdate";
 const armResourceUpdateRegex = "Azure\\.ResourceManager\\.@armResourceUpdate";
+
+export const extensionResourceOperationName = "@extensionResourceOperation";
+export const legacyExtensionResourceOperationName = "@legacyExtensionResourceOperation";
+export const legacyResourceOperationName = "@legacyResourceOperation";
+export const builtInResourceOperationName = "@builtInResourceOperation";
+
+export const armResourceWithParameter =
+  "Azure.ResourceManager.Private.@armResourceWithParameter";
+const armResourceWithParameterRegex =
+  "Azure\\.ResourceManager\\.Private\\.@armResourceWithParameter";
 
 export const armResourceInternal =
   "Azure.ResourceManager.Private.@armResourceInternal";
@@ -108,6 +120,7 @@ export const azureSDKContextOptions: CreateSdkContextOptions = {
     resourceGroupResourceRegex,
     singletonRegex,
     subscriptionResourceRegex,
-    tenantResourceRegex
+    tenantResourceRegex,
+    armResourceWithParameterRegex,
   ]
 };