+++ sdk/servicebus/Azure.Messaging.ServiceBus/samples/TopicFilters/Program.cs
@@ -27,18 +27,24 @@ namespace TopicSubscriptionWithRuleOperationsSample
 
         public static async Task Main(string[] args)
         {
+            var fullyQualifiedNamespaceOption = new Option<string>(
+                aliases: new[] { "--namespace" },
+                description: "Fully qualified Service Bus Queue namespace to use");
+
+            var connectionOption = new Option<string>(
+                aliases: new[] { "--connection-variable" },
+                description: "The name of an environment variable containing the connection string to use.");
+
             var command = new RootCommand("Demonstrates the Topic Filters feature of Azure Service Bus.")
             {
-                new Option<string>(
-                    alias: "--namespace",
-                    description: "Fully qualified Service Bus Queue namespace to use") {Name = "FullyQualifiedNamespace"},
-                new Option<string>(
-                    alias: "--connection-variable",
-                    description: "The name of an environment variable containing the connection string to use.") {Name = "Connection"},
+                fullyQualifiedNamespaceOption,
+                connectionOption
             };
-            command.Handler = CommandHandler.Create<string, string>(RunAsync);
+
+            command.SetHandler(RunAsync, fullyQualifiedNamespaceOption, connectionOption);
             await command.InvokeAsync(args);
         }
+
 
         private static async Task RunAsync(string fullyQualifiedNamespace, string connection)
         {
@@ -58,7 +64,7 @@ namespace TopicSubscriptionWithRuleOperationsSample
                 throw new ArgumentException(
                     "Either a fully qualified namespace or a connection string environment variable must be specified.");
             }
-            
+
             Console.WriteLine($"Creating topic {TopicName}");
             await s_adminClient.CreateTopicAsync(TopicName);
 
@@ -115,7 +121,7 @@ namespace TopicSubscriptionWithRuleOperationsSample
             // Send messages to Topic
             await SendMessagesAsync();
 
-            // Receive messages from 'NoFilterSubscription'. Should receive all 9 messages 
+            // Receive messages from 'NoFilterSubscription'. Should receive all 9 messages
             await ReceiveMessagesAsync(NoFilterSubscriptionName);
 
             // Receive messages from 'SqlFilterOnlySubscription'. Should receive all messages with Color = 'Red' i.e 3 messages
@@ -129,7 +135,7 @@ namespace TopicSubscriptionWithRuleOperationsSample
             // i.e 1 message
             await ReceiveMessagesAsync(CorrelationFilterSubscriptionName);
             Console.ResetColor();
-            
+
             Console.WriteLine("=======================================================================");
             Console.WriteLine("Completed Receiving all messages. Disposing clients and deleting topic.");
             Console.WriteLine("=======================================================================");
@@ -140,7 +146,7 @@ namespace TopicSubscriptionWithRuleOperationsSample
             await s_client.DisposeAsync();
 
             Console.WriteLine("Deleting topic");
-            
+
             // Deleting the topic will handle deleting all the subscriptions as well.
             await s_adminClient.DeleteTopicAsync(TopicName);
         }
@@ -186,7 +192,7 @@ namespace TopicSubscriptionWithRuleOperationsSample
             Console.WriteLine($"Created message with color: {message.ApplicationProperties["Color"]}, CorrelationId: {message.CorrelationId}");
             Console.ResetColor();
         }
-        
+
         private static void PrintReceivedMessage(ServiceBusReceivedMessage message)
         {
             Console.ForegroundColor = (ConsoleColor) Enum.Parse(typeof(ConsoleColor), message.Subject);