+++ sdk/core/System.ClientModel/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.9.0-beta.1 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+- Fixed an issue with `ClientRetryPolicy` where delays were being calculated using the retry count instead of the attempt count, causing the initial retry to occur without delay and subsequent retries to be performed more quickly than intended.
+
+### Other Changes
+
 ## 1.8.1 (2025-11-10)
 
 ### Bugs Fixed